Oid: Return to auto-derived traits
Rustc now seems perfectly capable of auto-deriving equality, ordering
and hashing traits for the Oid struct, so let it.  The hash really
doesn't need to look past the first 8-bytes or so, since the SHA-1
hash is, in theory, better than the hash used.  But, the speed of the
hash function is quite small compared to generating the SHA-1 in the
first place.

Also adds some benchmarks for putting the hashes in both a BTreeMap
and a HashMap.  As is similar to other platforms, the HashMap severely
outperforms the BTreeMap as the map grows larger.  It seems reasonable
to sort the results once the hash is finalized.