Update to latest rust/cargo master
Cargo and rust no longer support packages with '-' characters, and
make it appear to the code that existing ones have an underscore
there.

Rustc has a trivial cast warning, but hasn't implemented the intended
replacement (type ascription), so just disable the warning for now.

Cargo update to get upstream packages that will build as well.