cas: Use inner mutability on writer
Because the writer gets embedded in lots of other places, use inner
mutability so that this is actually possible.

For the sql-based pool, take advantage of the inner mutability of the
database connection itself.

TODO: This doesn't protect against multiple writers, at least not any
more than the underlying database does.