Write simple types to map to data/indirect blocks
The 'decode' module unpacks a chunk into a representation of the
backup node it represents.  Right now, implement "indirect" blocks,
and data blocks.

A consequence of this is that we now have enough of a public API to
encode and decode data, so move the unit tests out of the modules into
a proper test.  In order to be able to verify the data, we need to use
a deterministic prng to be able to check that the data is correct.

Jira: BACK-4