Fix up filer to work with new pool API.
Lift the sink parameter into various methods so we don't have to store
a mutable reference.  It's a little more complex to pass around, but
avoids having to manage the lifetime, and makes the indirect writer
possible, since its use is interleaved.

At this point, the filer tests are passing again.